pyproject.toml
[tool]
[tool.poetry]
name = "kube-downscaler"
version = "2020.1.1dev0"
description = "Scale down Kubernetes deployments after work hours"
authors = ["Henning Jacobs <henning@jacobs1.de>"]

[tool.poetry.dependencies]
python = ">=3.8,<4"
pykube-ng = "*"
pytz = "*"

[tool.poetry.dev-dependencies]
coverage = "*"
coveralls = "*"
flake8 = "*"
pytest = "*"
pytest-cov = "*"
black = "^22.3.0"
mypy = "^0.761"
pre-commit = "^2.9.2"
